demo/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
altair=4.2.2=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
augly=1.0.0=pypi_0
autopep8=1.6.0=pyhd3eb1b0_1
backports-zoneinfo=0.2.1=pypi_0
base58=2.1.1=pypi_0
bitarray=2.7.0=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
cryptography=38.0.4=py38h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cytoolz=0.12.1=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
eth-abi=2.2.0=pypi_0
eth-account=0.5.9=pypi_0
eth-hash=0.3.3=pypi_0
eth-keyfile=0.5.1=pypi_0
eth-keys=0.3.4=pypi_0
eth-rlp=0.2.1=pypi_0
eth-typing=2.3.0=pypi_0
eth-utils=1.10.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
giflib=5.2.1=h5eee18b_1
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
hamming-codec=0.3.5=pypi_0
hexbytes=0.3.0=pypi_0
huggingface-hub=0.12.0=pypi_0
idna=3.4=py38h06a4308_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.10=pypi_0
ipfshttpclient=0.8.0a2=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h7f8727e_0
jsonschema=4.17.3=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lru-dict=1.1.8=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=2.1.0=pypi_0
markupsafe=2.1.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multiaddr=0.0.9=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
netaddr=0.8.0=pypi_0
nettle=3.7.3=hbbd107a_1
numpy=1.24.2=pypi_0
numpy-base=1.23.5=py38h31eccc5_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
parsimonious=0.8.1=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
portalocker=2.7.0=pypi_0
protobuf=3.19.5=pypi_0
pyarrow=11.0.0=pypi_0
pycodestyle=2.10.0=py38h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.17=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.15=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-magic=0.4.27=pypi_0
pytorch=1.12.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
rich=13.3.1=pypi_0
rlp=2.0.1=pypi_0
scipy=1.10.0=pypi_0
semver=2.13.0=pypi_0
setuptools=67.2.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
streamlit=1.17.0=pypi_0
timm=0.6.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
torch=1.13.1=pypi_0
torchaudio=0.12.1=py38_cu102
torchvision=0.14.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
typer=0.7.0=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.14=py38h06a4308_0
validators=0.20.0=pypi_0
varint=1.0.2=pypi_0
watchdog=2.2.1=pypi_0
web3=5.31.3=pypi_0
websockets=9.1=pypi_0
wheel=0.38.4=pypi_0
xz=5.2.10=h5eee18b_1
yarl=1.8.2=pypi_0
zipp=3.12.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
